GLVersionTest Passes, need to re-visit GLWindowSizeTest
After reviewing the GLFW documentation, simplified how the major and
minor version numbers are handled and use the glfwWindowHint and
glfwGetWindowAttrib functions to handle this instead. Also added some
error checking functions for the version and sizes.

Will plan to create a CreateWindow method for the default constructor
that will do what the GLWindow constructor does for added flexibility.