Redid setting of size and title
Used the glfw methods to set the size and title for a GLWindow class. A
title variable is still needed since there is no "glfwGetWindowTitle".
Also, decorators have been turned off by default as this affects the
minimum size for the window. Will allow methods in the future to modify
this.