Added New Shader Test for Buffer Attachment. Modified GLWindow Test.
Added GLAttachBufferToShaderTest which tests that a GLBuffer can
successfully be attached to a GLShader.

Removed the the call to glfwWindowHint in GLWindow::Create which removes
any window decorations as it is cumbersome to not be able to move a
window. Modified the GLWindowCreationTest to reflect this so it still
passes successfully.