Simplified GLModifyUniformsTest
After realizing that GLSL optimizes out unused uniforms, the vertex
shader used only uses 3 uniform variables to test with. The methodology
is that the function calls are similar enough that if the calls work for
vec4, an array of vec4, and a mat4, it should work for all other uniform
calls.

All tests pass.